# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: "1.0.0"
  description: >
    This document defines the API for the WASAText messaging application.
    It includes endpoints for user login, updating user details,
    managing conversations, sending messages, and handling group operations.
servers:
  - url: https://localhost:3000
    description: Local development server
paths:
...
  /message:
    post:
      tags:
        - messages
      summary: Send a message
      description: >
        Sends a new message. For private chats, if no conversation exists, one is auto-created.
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - isGroup
              properties:
                receiverId:
                  type: string
                  format: uuid
                  description: Required for private messages
                  example: "receiver-uuid"
                content:
                  type: string
                  minLength: 1
                  maxLength: 5000
                  example: "Hello, world!"
                isGroup:
                  type: boolean
                  example: false
                groupId:
                  type: string
                  format: uuid
                  description: Required when isGroup is true
                  example: "group-uuid"
      responses:
        '201':
          description: Message sent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageId:
                    type: string
                    format: uuid
                    example: "message-uuid"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /message/{messageId}/forward:
    post:
      tags:
        - messages
      summary: Forward a message
      description: Forwards an existing message to another conversation.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the message to forward.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetConversationId
              properties:
                targetConversationId:
                  type: string
                  format: uuid
                  example: "target-conversation-uuid"
      responses:
        '200':
          description: Message forwarded successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageId:
                    type: string
                    format: uuid
                    example: "new-forwarded-message-uuid"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /message/{messageId}/comment:
    post:
      tags:
        - messages
      summary: Comment on a message
      description: Adds a reaction (comment) to a message.
      operationId: commentMessage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the message to comment on.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reaction
              properties:
                reaction:
                  type: string
                  minLength: 1
                  maxLength: 5
                  example: "👍"
      responses:
        '201':
          description: Comment added successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /message/{messageId}/delete:
    delete:
      tags:
        - messages
      summary: Delete a message
      description: Deletes a message if the authenticated user is the sender.
      operationId: deleteMessage
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the message to delete.
      responses:
        '200':
          description: Message deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/add:
    post:
      tags:
        - groups
      summary: Add a user to a group
      description: Adds the authenticated user to the specified group.
      operationId: addToGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - groupId
              properties:
                groupId:
                  type: string
                  format: uuid
                  example: "group-uuid"
      responses:
        '200':
          description: User added to group successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/leave:
    delete:
      tags:
        - groups
      summary: Leave a group
      description: Removes the authenticated user from the specified group.
      operationId: leaveGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - groupId
              properties:
                groupId:
                  type: string
                  format: uuid
                  example: "group-uuid"
      responses:
        '200':
          description: Left group successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/name:
    put:
      tags:
        - groups
      summary: Update group name
      description: Updates the name of a group.
      operationId: setGroupName
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - groupId
                - newName
              properties:
                groupId:
                  type: string
                  format: uuid
                  example: "group-uuid"
                newName:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: "^[\\w\\s-]+$"
                  example: "New Group Name"
      responses:
        '200':
          description: Group name updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/photo:
    put:
      tags:
        - groups
      summary: Update group photo
      description: Updates the photo URL of a group.
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - groupId
                - photoUrl
              properties:
                groupId:
                  type: string
                  format: uuid
                  example: "group-uuid"
                photoUrl:
                  type: string
                  format: uri
                  pattern: "^https://.+$"
                  example: "https://example.com/newgroupphoto.jpg"
      responses:
        '200':
          description: Group photo updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
          
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token obtained from /session endpoint.
        Token should be included in the Authorization header as "Bearer <token>".
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid request parameters"
        details:
          type: array
          items:
            type: string
          example: ["Field 'name' must be between 3 and 16 characters"]
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'